[WIP] Protractor and Jasmine infra
Basic infrastructure for Protractor automation and the Jasmine BDD
framework. To run it:

yarn install
./node_modules/protractor/bin/webdriver-manager update
yarn test-suite --suite base

You will see a browser briefly open up. On the console, you will see that
a test for "GUI layout" was run, and that it tested that it "includes the
Cluster Name field".